requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
lxml==4.6.1
py==1.9.0
pymongo==3.11.0
pyshark==0.4.2.11
pytz==2020.4
PyYAML==5.3.1
ruamel.yaml==0.16.12       
ruamel.yaml.clib==0.2.2    
sqlparse==0.4.1
